+++ OKX.Api/Account/Responses/OkxAccountLeverage.cs
@@ -27,5 +27,5 @@ public record OkxAccountLeverage
     /// Leverage
     /// </summary>
     [JsonProperty("lever")]
-    public int Leverage { get; set; }
+    public decimal Leverage { get; set; }
 }
