+++ OKX.Api/CopyTrading/Responses/OkxCopyTradingLeadTraderPosition.cs
@@ -33,7 +33,7 @@ public record OkxCopyTradingLeadTraderPosition
     /// Leverage
     /// </summary>
     [JsonProperty("lever")]
-    public int Leverage { get; set; }
+    public decimal Leverage { get; set; }
 
     /// <summary>
     /// Average open price
