+++ OKX.Api/SignalBot/Responses/OkxSignalBotPositionHistory.cs
@@ -69,7 +69,7 @@ public record OkxSignalBotPositionHistory
     /// Leverage
     /// </summary>
     [JsonProperty("lever")]
-    public int Leverage { get; set; }
+    public decimal Leverage { get; set; }
     
     /// <summary>
     /// Direction: long short
